Simon domasno 3 location functionalities

+++ Domasna3/src/main/java/com/example/diansspring/model/Facility.java
@@ -30,7 +30,7 @@ public class Facility {
     @OneToMany(mappedBy = "facility")
     private List<Review> reviews;
 
-    private float reviewRatingsSum;
+    private double reviewRatingsSum;
 
     private int reviewRatingsCount;
